build 1.1.7 : some changes, including change of toktype to int64 to take advantage of go 1.19 switch jump tables.

+++ constants.go
@@ -164,7 +164,7 @@ const (
 //  a few of these are unused now and some should probably be renamed.
 //  they should be checked and tidied next time there is any change to be done here.
 const (
-	Error uint8 = iota
+	Error int64 = iota
 	EOL
 	EOF
 	StringLiteral
